Renaming files and variables for Surmise
